+++ random.h
@@ -1,6 +1,6 @@
 #ifndef RANDOM
 #define RANDOM
 
-unsigned long int getRandomNumber(int n);
+unsigned long int getRandomNumber(unsigned long int n);
 
 #endif
